refactor(core): consolidate layout geometric types into models
The layout modules had grown their own local copies of types that
already existed (or should exist) in `core/models/types.ts`:

  - `Position`   defined in both `models/types.ts` and `sugiyama/coords.ts`
  - `Bounds`     defined in both `models/types.ts` and `sugiyama/compound.ts`
  - `Direction`  string union redeclared in `coords.ts`, `port-placement.ts`,
                 `network-layout.ts` (plus an orphan `LayoutDirection` alias)
  - `NodeSize`   local alias for what is really `Size` in models

Consolidating:

  - Promote `Direction` to `core/models/types.ts`; remove the orphan
    `LayoutDirection` (no external consumers — the two internal uses
    had just renamed to `Direction`).
  - Use the existing `Size` from `models/types.ts` everywhere in place
    of the ad-hoc `NodeSize`.
  - Delete the duplicate `Position` interface from `coords.ts` and the
    duplicate `Bounds` from `compound.ts`; re-import from core.
  - Adjacent files (`port-placement.ts`, `network-layout.ts`) also
    switched from inline string unions to the shared `Direction` type.

No runtime behaviour change — types are structurally identical to
what was there before, just sourced from one place. Public API shapes
(`layoutNetwork(graph, options)`, `placeNode(…)`, etc.) are unchanged.

Tests: 116/116 core pass, 30/30 workspace typecheck.

Part of the layout-API cleanup series: step 1 (types), step 2 (option
types), step 3 (hints + two-API docs).
