+++ libs/@shumoku/core/src/layout/sugiyama/index.ts
@@ -4,29 +4,25 @@
 /**
  * Sugiyama-style layered layout pipeline.
  *
- * This module is a parallel implementation of `layoutNetwork`, built out
- * phase by phase. It is not wired into the public `layoutNetwork` yet —
- * when the four phases are complete and validated against the existing
- * tree-based output, we will switch `layoutNetwork` to delegate here
- * and remove the legacy implementation.
+ * `layoutNetwork` delegates its internals to `layoutCompound` here;
+ * callers outside core rarely need these primitives directly, but they
+ * are exported for testability and so other layout-engine experiments
+ * can reuse the individual phases without vendoring the whole module.
+ *
+ * Shared geometric types (`Position`, `Bounds`, `Size`, `Direction`)
+ * live in `@shumoku/core/models` rather than being redefined here.
  */
 
 export type { LayoutFlatOptions, LayoutFlatResult } from './compose.js'
 export { layoutFlat } from './compose.js'
 export type {
-  Bounds,
   CompoundLayoutOptions,
   CompoundLayoutResult,
   CompoundNode,
   CompoundSubgraph,
 } from './compound.js'
 export { layoutCompound } from './compound.js'
-export type {
-  AssignCoordinatesOptions,
-  Direction,
-  NodeSize,
-  Position,
-} from './coords.js'
+export type { AssignCoordinatesOptions } from './coords.js'
 export { assignCoordinates } from './coords.js'
 export { removeCycles } from './cycles.js'
 export { assignLayers } from './layers.js'
